Split base contract and abstract contract
