+++ skale/contracts/config_controller/config_controller.py
@@ -20,7 +20,8 @@
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.abstract_contract import transaction_method
+from skale.contracts.base_contract import BaseContract
 
 
 class ConfigController(BaseContract):
