+++ skale/contracts/context/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.base_contract import BaseContract
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.context.context import Context
 
 __all__ = ['Context']
