+++ skale/contracts/fair/committee.py
@@ -21,7 +21,8 @@ from typing import Any
 
 from eth_typing import ChecksumAddress
 
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.abstract_contract import transaction_method
+from skale.contracts.base_contract import BaseContract
 from skale.types.committee import Committee as CommitteeStruct
 from skale.types.committee import CommitteeIndex
 from skale.types.node import NodeId
