+++ skale/contracts/fair/staking.py
@@ -21,7 +21,8 @@ from typing import List
 
 from eth_typing import ChecksumAddress
 
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.abstract_contract import transaction_method
+from skale.contracts.base_contract import BaseContract
 from skale.types.exit_request import ExitRequest, exit_request_from_tuple
 from skale.types.node import NodeId
