+++ skale/contracts/ima/__init__.py
@@ -1,4 +1,5 @@
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.abstract_contract import transaction_method
+from skale.contracts.base_contract import BaseContract
 from skale.contracts.ima.mainnet.linker import Linker
 from skale.contracts.manager.contract_manager import ContractManager
