+++ skale/contracts/ima/schain/token_manager_erc721.py
@@ -21,7 +21,8 @@ from Crypto.Hash import keccak
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import BaseContract, transaction_method
+from skale.contracts.abstract_contract import transaction_method
+from skale.contracts.base_contract import BaseContract
 from skale.types.schain import SchainName
