+++ skale/contracts/manager/constants_holder.py
@@ -20,7 +20,7 @@
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
