+++ skale/contracts/manager/contract_manager.py
@@ -23,7 +23,7 @@ from eth_typing import ChecksumAddress
 from web3 import Web3
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.utils.helper import add_0x_prefix
