+++ skale/contracts/manager/delegation/delegation_controller.py
@@ -23,7 +23,7 @@ from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 from web3.types import Wei
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.types.delegation import Delegation, DelegationId, DelegationStatus, FullDelegation
 from skale.types.validator import ValidatorId
