+++ skale/contracts/manager/delegation/validator_service.py
@@ -24,7 +24,7 @@ from web3 import Web3
 from web3.contract.contract import ContractFunction
 from web3.types import HexBytes, Wei
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.types.validator import Validator, ValidatorId, ValidatorWithId
 from skale.utils.helper import format_fields
