+++ skale/contracts/manager/dkg.py
@@ -22,7 +22,7 @@ from typing import List, Tuple
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.transactions.tools import retry_tx
 from skale.types.dkg import G2Point, KeyShare, VerificationVector
