+++ skale/contracts/manager/manager.py
@@ -25,7 +25,7 @@ from eth_abi.abi import encode
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.dataclasses.schain_options import SchainOptions, get_default_schain_options
 from skale.transactions.result import TxRes
