+++ skale/contracts/manager/nodes.py
@@ -26,7 +26,7 @@ from eth_typing import BlockNumber, ChecksumAddress, HexStr
 from web3.contract.contract import ContractFunction
 from web3.exceptions import BadFunctionCallOutput, ContractLogicError
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.types.node import Node, NodeId, NodeStatus, Port
 from skale.types.validator import ValidatorId
