+++ skale/contracts/manager/schains.py
@@ -29,7 +29,7 @@ from web3 import Web3
 from web3.contract.contract import ContractFunction
 from web3.types import Wei
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.manager.node_rotation import NodeRotation
 from skale.contracts.manager.schains_internal import SChainsInternal
 from skale.contracts.skale_manager_contract import SkaleManagerContract
