+++ skale/contracts/manager/sync_manager.py
@@ -24,7 +24,7 @@ from typing import List
 from eth_typing import ChecksumAddress
 from web3.contract.contract import ContractFunction
 
-from skale.contracts.base_contract import transaction_method
+from skale.contracts.abstract_contract import transaction_method
 from skale.contracts.skale_manager_contract import SkaleManagerContract
 from skale.transactions.result import TxRes
 from skale.utils.helper import ip_from_bytes, ip_to_bytes
