+++ skale/transactions/tools.py
@@ -21,7 +21,7 @@
 import logging
 import time
 from functools import partial, wraps
-from typing import TYPE_CHECKING, Any, Callable, Optional
+from typing import Any, Callable, Optional
 
 from eth_typing import ChecksumAddress
 from web3 import Web3
@@ -34,10 +34,7 @@ import skale.config as config
 from skale.transactions.exceptions import TransactionError
 from skale.transactions.result import TxCallResult, TxRes, TxStatus
 from skale.utils.web3_utils import get_eth_nonce
-
-if TYPE_CHECKING:
-    from skale.skale_base import SkaleBase
-
+from skale.wallets import BaseWallet
 
 logger = logging.getLogger(__name__)
 
@@ -46,13 +43,17 @@ DEFAULT_ETH_SEND_GAS_LIMIT = 22000
 
 
 def make_dry_run_call(
-    skale: 'SkaleBase', method: ContractFunction, gas_limit: int | None = None, value: Wei = Wei(0)
+    web3: Web3,
+    wallet: BaseWallet,
+    method: ContractFunction,
+    gas_limit: int | None = None,
+    value: Wei = Wei(0),
 ) -> TxCallResult:
-    opts = TxParams({'from': skale.wallet.address, 'value': value})
+    opts = TxParams({'from': wallet.address, 'value': value})
     logger.info(
         f'Dry run tx: {method.fn_name}, '
-        f'sender: {skale.wallet.address}, '
-        f'wallet: {skale.wallet.__class__.__name__}, '
+        f'sender: {wallet.address}, '
+        f'wallet: {wallet.__class__.__name__}, '
         f'value: {value}, '
     )
     estimated_gas = 0
@@ -63,7 +64,7 @@ def make_dry_run_call(
             opts.update({'gas': gas_limit})
             method.call(opts)
         else:
-            estimated_gas = estimate_gas(skale.web3, method, opts)
+            estimated_gas = estimate_gas(web3, method, opts)
         logger.info(f'Estimated gas for {method.fn_name}: {estimated_gas}')
     except ContractLogicError as e:
         message = e.message or 'Contract logic error'
